+++ project/src/const.tsx
@@ -1,0 +1,6 @@
+export enum AppRoute {
+  SignIn = '/login',
+  Main = '/',
+  Favorites = '/favorites',
+  Room = '/offer/:id'
+}
