+++ project/src/types/types.tsx
@@ -1,0 +1,5 @@
+type AppProps = {
+  countCard: number,
+};
+
+export default AppProps;
